Move get_amount_from_tx_env to currency crate
