+++ pallets/issue/src/ext.rs
@@ -2,8 +2,29 @@
 use mocktopus::macros::mockable;
 
 #[cfg_attr(test, mockable)]
+pub(crate) mod currency {
+	use substrate_stellar_sdk::TransactionEnvelope;
+
+	use currency::{Amount, Error};
+	use primitives::StellarPublicKeyRaw;
+
+	use crate::types::CurrencyId;
+
+	pub fn get_amount_from_transaction_envelope<T: crate::Config>(
+		transaction_envelope: &TransactionEnvelope,
+		recipient_stellar_address: StellarPublicKeyRaw,
+		currency: CurrencyId<T>,
+	) -> Result<Amount<T>, Error<T>> {
+		<currency::Pallet<T>>::get_amount_from_transaction_envelope(
+			transaction_envelope,
+			recipient_stellar_address,
+			currency,
+		)
+	}
+}
+
+#[cfg_attr(test, mockable)]
 pub(crate) mod stellar_relay {
-	use currency::Amount;
 	use sp_std::convert::TryFrom;
 	use substrate_stellar_sdk::{
 		compound_types::UnlimitedVarArray,
@@ -11,6 +32,7 @@ pub(crate) mod stellar_relay {
 		TransactionEnvelope, XdrCodec,
 	};
 
+	use currency::Amount;
 	use primitives::StellarPublicKeyRaw;
 	use stellar_relay::Error;
 
@@ -32,18 +54,6 @@ pub(crate) mod stellar_relay {
 		raw_encoded_xdr: &[u8],
 	) -> Result<V, Error<T>> {
 		<stellar_relay::Pallet<T>>::construct_from_raw_encoded_xdr(raw_encoded_xdr)
-	}
-
-	pub fn get_amount_from_transaction_envelope<T: crate::Config, V: TryFrom<i64>>(
-		transaction_envelope: &TransactionEnvelope,
-		recipient_stellar_address: StellarPublicKeyRaw,
-		currency: CurrencyId<T>,
-	) -> Result<Amount<T>, currency::Error<T>> {
-		<stellar_relay::Pallet<T>>::get_amount_from_transaction_envelope(
-			transaction_envelope,
-			recipient_stellar_address,
-			currency,
-		)
 	}
 }
