+++ pallets/stellar-relay/src/lib.rs
@@ -548,98 +548,6 @@ pub mod pallet {
 			let decoded = V::from_xdr(value_xdr).map_err(|_| Error::<T>::InvalidXDR)?;
 			Ok(decoded)
 		}
-
-		/// Accumulate the amounts of the specified currency that happened in the operations of a
-		/// Stellar transaction
-		pub fn get_amount_from_transaction_envelope(
-			transaction_envelope: &TransactionEnvelope,
-			recipient_stellar_address: Uint256,
-			currency: CurrencyId<T>,
-		) -> Result<Amount<T>, currency::Error<T>> {
-			let asset = AssetConversionOf::<T>::lookup(currency)
-				.map_err(|_| currency::Error::<T>::AssetConversionError.into())?;
-			let recipient_account_muxed = MuxedAccount::KeyTypeEd25519(recipient_stellar_address);
-			let recipient_account_pk = PublicKey::PublicKeyTypeEd25519(recipient_stellar_address);
-
-			let amount: i64 = match transaction_envelope {
-				TransactionEnvelope::EnvelopeTypeTxV0(envelope) => {
-					let mut sum: i64 = 0;
-					for x in envelope.tx.operations.get_vec().iter() {
-						match x.body.clone() {
-							OperationBody::Payment(payment) => {
-								if payment.destination.eq(&recipient_account_muxed) &&
-									payment.asset == asset
-								{
-									sum = sum.saturating_add(payment.amount);
-								}
-							},
-							OperationBody::CreateClaimableBalance(payment) => {
-								// for security reasons, we only count operations that have the
-								// recipient as a single claimer and unconditional claim predicate
-								if payment.claimants.len() == 1 {
-									if let Claimant::ClaimantTypeV0(claimant) =
-										payment.claimants.get_vec()[0].clone()
-									{
-										if claimant.destination.eq(&recipient_account_pk) &&
-											claimant.predicate ==
-												ClaimPredicate::ClaimPredicateUnconditional
-										{
-											sum = sum.saturating_add(payment.amount);
-										}
-									}
-								}
-							},
-							_ => {
-								// ignore other operations
-							},
-						}
-					}
-					// return the sum of the amounts
-					sum
-				},
-				TransactionEnvelope::EnvelopeTypeTx(envelope) => {
-					let mut sum: i64 = 0;
-					for x in envelope.tx.operations.get_vec().iter() {
-						match x.body.clone() {
-							OperationBody::Payment(payment) => {
-								if payment.destination.eq(&recipient_account_muxed) &&
-									payment.asset == asset
-								{
-									sum = sum.saturating_add(payment.amount);
-								}
-							},
-							OperationBody::CreateClaimableBalance(payment) => {
-								// for security reasons, we only count operations that have the
-								// recipient as a single claimer and unconditional claim predicate
-								if payment.claimants.len() == 1 {
-									if let Claimant::ClaimantTypeV0(claimant) =
-										payment.claimants.get_vec()[0].clone()
-									{
-										if claimant.destination.eq(&recipient_account_pk) &&
-											claimant.predicate ==
-												ClaimPredicate::ClaimPredicateUnconditional
-										{
-											sum = sum.saturating_add(payment.amount);
-										}
-									}
-								}
-							},
-							_ => {
-								// ignore other operations
-							},
-						}
-					}
-					// return the sum of the amounts
-					sum
-				},
-				TransactionEnvelope::EnvelopeTypeTxFeeBump(_) => 0,
-				TransactionEnvelope::Default(_) => 0,
-			};
-
-			let balance = BalanceConversionOf::<T>::unlookup(amount);
-			let amount: Amount<T> = Amount::new(balance, currency);
-			Ok(amount)
-		}
 	}
 
 	pub fn compute_non_generic_tx_set_content_hash(tx_set: &TransactionSet) -> [u8; 32] {
